infra/cloudfunctions/requirements.txt
google-cloud-pubsub
google-cloud-container
google-cloud-build
google-cloud-firestore